requirements.txt
python-telegram-bot==20.8
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==10.2.0
Flask==3.0.2
gunicorn==21.2.0
